# expression: go generate vectorized addtime functions