--- a/.codecov.yml
+++ b/.codecov.yml
@@ -42,4 +42,5 @@ ignore:
   - "ddl/testutil/.*"
   - "executor/seqtest/.*"
   - "metrics/.*"
+  - "expression/generator/.*"