--- a/expression/builtin_cast_vec_test.go
+++ b/expression/builtin_cast_vec_test.go
@@ -28,7 +28,27 @@ var vecBuiltinCastCases = map[string][]vecExprBenchCase{
 		{retEvalType: types.ETDuration, childrenTypes: []types.EvalType{types.ETInt}, geners: []dataGenerator{new(randDurInt)}},
 		{retEvalType: types.ETReal, childrenTypes: []types.EvalType{types.ETReal}},
 		{retEvalType: types.ETReal, childrenTypes: []types.EvalType{types.ETDecimal}},
+		{retEvalType: types.ETDuration, childrenTypes: []types.EvalType{types.ETDatetime},
+			geners: []dataGenerator{&dateTimeGenerWithFsp{
+				defaultGener: defaultGener{nullRation: 0.2, eType: types.ETDatetime},
+				fsp:          1,
+			}},
+		},
 	},
+}
+
+type dateTimeGenerWithFsp struct {
+	defaultGener
+	fsp int8
+}
+
+func (g *dateTimeGenerWithFsp) gen() interface{} {
+	result := g.defaultGener.gen()
+	if t, ok := result.(types.Time); ok {
+		t.Fsp = g.fsp
+		return t
+	}
+	return result
 }
 
 func (s *testEvaluatorSuite) TestVectorizedBuiltinCastEvalOneVec(c *C) {