--- a/types/overflow_test.go
+++ b/types/overflow_test.go
@@ -15,6 +15,7 @@ package types
 
 import (
 	"math"
+	"time"
 
 	. "github.com/pingcap/check"
 	"github.com/pingcap/tidb/util/testleak"
@@ -68,6 +69,12 @@ func (s *testOverflowSuite) TestAdd(c *C) {
 			c.Assert(err, NotNil)
 		} else {
 			c.Assert(ret, Equals, t.ret)
+		}
+		ret2, err := AddDuration(time.Duration(t.lsh), time.Duration(t.rsh))
+		if t.overflow {
+			c.Assert(err, NotNil)
+		} else {
+			c.Assert(ret2, Equals, time.Duration(t.ret))
 		}
 	}