<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Use go template to generate vectorized [ADDTIME](https://dev.mysql.com/doc/refman/5.7/en/date-and-time-functions.html#function_addtime) functions.
* builtinAddDatetimeAndDurationSig
* builtinAddDatetimeAndStringSig
* builtinAddDurationAndDurationSig
* builtinAddDurationAndStringSig
* builtinAddStringAndDurationSig
* builtinAddStringAndStringSig
* builtinAddDateAndDurationSig
* builtinAddDateAndStringSig

By hand write a cast function for unit test.
* builtinCastTimeAsDurationSig


### What is changed and how it works?

* addtime
```
name                                                                         old time/op    new time/op    delta
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDatetimeAndDurationSig12     174µs ±16%      62µs ± 4%   -64.28%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDatetimeAndStringSig12      1.37ms ± 4%    1.19ms ± 7%   -13.15%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDurationAndDurationSig12    74.5µs ±79%     5.3µs ± 9%   -92.89%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDurationAndStringSig12      1.17ms ± 3%    1.10ms ± 3%    -5.56%  (p=0.016 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddStringAndDurationSig12      1.48ms ± 2%    1.41ms ± 3%    -4.83%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddStringAndStringSig12        2.33ms ± 3%    2.25ms ± 3%      ~     (p=0.056 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDateAndDurationSig12         409µs ±14%     341µs ± 0%   -16.66%  (p=0.016 n=5+4)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDateAndStringSig12          1.73ms ± 4%    1.68ms ± 6%      ~     (p=0.310 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeDateTimeNullSig12        142µs ±68%       0µs ± 3%   -99.95%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeStringNullSig12         43.0µs ±42%     2.1µs ± 3%   -95.10%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeDurationNullSig12       22.5µs ±11%     0.1µs ± 3%   -99.69%  (p=0.016 n=4+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDatetimeAndDurationSig12           100µs ± 5%      62µs ± 4%   -37.26%  (p=0.008 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDatetimeAndStringSig12            1.47ms ± 3%    1.41ms ± 1%      ~     (p=0.056 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDurationAndDurationSig12          29.7µs ± 2%     4.8µs ± 5%   -83.69%  (p=0.008 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDurationAndStringSig12            1.18ms ± 7%    1.12ms ± 8%      ~     (p=0.095 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndDurationSig12            1.53ms ± 5%    1.57ms ± 8%      ~     (p=0.841 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndStringSig12              2.46ms ± 3%    2.43ms ± 6%      ~     (p=0.690 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDateAndDurationSig12               352µs ± 6%     315µs ±10%   -10.59%  (p=0.032 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDateAndStringSig12                1.58ms ± 4%    1.66ms ±11%      ~     (p=0.421 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeDateTimeNullSig12             12.3µs ± 3%     0.1µs ± 3%   -99.48%  (p=0.008 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeStringNullSig12               5.14µs ± 3%    2.02µs ± 5%   -60.80%  (p=0.008 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeDurationNullSig12             7.57µs ± 2%    0.06µs ± 3%   -99.17%  (p=0.008 n=5+5)

name                                                                         old alloc/op   new alloc/op   delta
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDatetimeAndDurationSig12     106kB ±24%       0kB       -100.00%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDatetimeAndStringSig12       210kB ±11%      96kB ± 0%   -54.31%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDurationAndDurationSig12   65.5kB ±136%     0.0kB       -100.00%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDurationAndStringSig12       147kB ± 8%      97kB ± 0%   -34.12%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddStringAndDurationSig12       271kB ±10%     190kB ± 0%   -29.88%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddStringAndStringSig12         386kB ± 5%     292kB ± 0%   -24.28%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDateAndDurationSig12         189kB ± 8%     101kB ± 0%   -46.54%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDateAndStringSig12           318kB ± 5%     212kB ± 0%   -33.45%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeDateTimeNullSig12        126kB ±36%       0kB       -100.00%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeStringNullSig12         44.4kB ±22%     0.0kB       -100.00%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeDurationNullSig12      39.3kB ±115%     0.0kB       -100.00%  (p=0.008 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDatetimeAndDurationSig12           0.00B          0.00B           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDatetimeAndStringSig12             102kB ± 0%     102kB ± 0%      ~     (p=0.841 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDurationAndDurationSig12           0.00B          0.00B           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDurationAndStringSig12            96.2kB ± 0%    96.4kB ± 0%      ~     (p=0.690 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndDurationSig12             198kB ± 0%     198kB ± 0%      ~     (p=0.310 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndStringSig12               290kB ± 0%     290kB ± 0%      ~     (p=0.841 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddDateAndDurationSig12              98.2kB ± 0%    98.2kB ± 0%    -0.00%  (p=0.000 n=5+4)
VectorizedBuiltinTimeFuncGenerated/builtinAddDateAndStringSig12                 211kB ± 0%     210kB ± 0%      ~     (p=0.222 n=5+5)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeDateTimeNullSig12              0.00B          0.00B           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeStringNullSig12                0.00B          0.00B           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeDurationNullSig12              0.00B          0.00B           ~     (all equal)

name                                                                         old allocs/op  new allocs/op  delta
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDatetimeAndDurationSig12      0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDatetimeAndStringSig12       5.82k ± 0%     5.82k ± 0%      ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDurationAndDurationSig12      0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDurationAndStringSig12       5.88k ± 0%     5.88k ± 0%      ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddStringAndDurationSig12       9.26k ± 0%     9.26k ± 0%    +0.01%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddStringAndStringSig12         15.3k ± 0%     15.3k ± 0%    +0.01%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDateAndDurationSig12         3.96k ± 0%     3.96k ± 0%      ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddDateAndStringSig12           10.5k ± 0%     10.5k ± 0%    +0.01%  (p=0.008 n=5+5)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeDateTimeNullSig12         0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeStringNullSig12           0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeEvalOneVecGenerated/builtinAddTimeDurationNullSig12         0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDatetimeAndDurationSig12            0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDatetimeAndStringSig12             6.18k ± 0%     6.18k ± 0%      ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDurationAndDurationSig12            0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDurationAndStringSig12             5.84k ± 0%     5.84k ± 0%      ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndDurationSig12             9.65k ± 0%     9.65k ± 0%      ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndStringSig12               15.1k ± 0%     15.1k ± 0%      ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDateAndDurationSig12               3.86k ± 0%     3.86k ± 0%      ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddDateAndStringSig12                 10.4k ± 0%     10.4k ± 0%      ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeDateTimeNullSig12               0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeStringNullSig12                 0.00           0.00           ~     (all equal)
VectorizedBuiltinTimeFuncGenerated/builtinAddTimeDurationNullSig12               0.00           0.00           ~     (all equal)
```

* cast: 
```
name                                                        old time/op    new time/op    delta
VectorizedBuiltinCastFunc/builtinCastTimeAsDurationSig--12    48.7µs ± 4%    32.0µs ± 4%  -34.27%  (p=0.008 n=5+5)

name                                                        old alloc/op   new alloc/op   delta
VectorizedBuiltinCastFunc/builtinCastTimeAsDurationSig--12     0.00B          0.00B          ~     (all equal)

name                                                        old allocs/op  new allocs/op  delta
VectorizedBuiltinCastFunc/builtinCastTimeAsDurationSig--12      0.00           0.00          ~     (all equal)
```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test