@qw4990 2019-09-18
> > in the performance benchmarks for add_time:
> > ```
> > name                                                                     old time/op    new time/op     delta
> > VectorizedBuiltinTimeFuncGenerated/builtinAddTimeStringNullSig--12         2.98µs ± 4%    77.28µs ±91%  +2494.20%  (p=0.008 n=5+5)
> > ```
> > 
> > 
> > seems there is a performance regression?
> 
> I think it's the problem of the benchmark framework.
> `vecEvalString` needs to apply its results into a `*chunk.Column` in benchmark`VecBuiltinFunc`, meanwhile `evalString` does not need in `NonVecBuiltinFunc`

I will fix it soon.