@SunRunAway 2019-09-24
> > ```
> > VectorizedBuiltinTimeFuncGenerated/builtinAddStringAndStringSig12              2.72ms ±10%    3.25ms ±12%   +19.49%  (p=0.032 n=5+5)
> > ```
> > 
> > 
> > Why is there performance regression?
> 
> I've investigated a little, but have no idea right now. :(

It has some difference from `builtinAddStringAndStringSig.evalString` and `builtinAddStringAndStringSig.vecEvalString`, but each of them has bug, see #12342.
I'll try to fix it in #12342