@SunRunAway 2019-09-26
> Almost LGTM, but please resolve conflicts. @SunRunAway

Rebased, conflicts resolved in builtin_cast_vec.go, builtin_cast_vec_test.go and builtin_time_vec_test.go